Small fixes

+++ crates/l2/monitor/widget/blocks.rs
@@ -1,7 +1,6 @@
 use std::cmp::min;
 
 use ethrex_common::{Address, H256, types::Block};
-use ethrex_rlp::encode::RLPEncode;
 use ethrex_storage::Store;
 use ratatui::{
     buffer::Buffer,
@@ -113,7 +112,7 @@ impl BlocksTable {
                     block.header.coinbase,
                     block.header.gas_used,
                     block.header.blob_gas_used,
-                    block.get_rlp_encode_size(),
+                    block.get_rlp_encode_size() as usize,
                 )
             })
             .collect::<Vec<_>>();
